requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
numpyy